ohacms/api/views.py
from django.contrib.auth.models import Post, Option
from rest_framework import viewsets
from rest_framework import permissions
from ohacms.serializers import PostSerializer, OptionSerializer


class PostViewSet(viewsets.ModelViewSet):
    """
    API endpoint that allows users to be viewed or edited.
    """
    queryset = Post.objects.all()
    serializer_class = PostSerializer
    #permission_classes = [permissions.Is...]


class OptionViewSet(viewsets.ModelViewSet):
    """
    API endpoint that allows groups to be viewed or edited.
    """
    queryset = Option.objects.all()
    serializer_class = OptionSerializer
    permission_classes = [permissions.IsAuthenticated]
